+++ R/adlb.R
@@ -3,7 +3,7 @@
 #' @description adlb modified from pharmaverseadam
 #' @source data from pharmaverseadam.
 #'
-#' @format A data frame with 83640 rows and 156 variables:
+#' @format A data frame with 83640 rows and 153 variables:
 #' \describe{
 #'  \item{STUDYID}{Study Identifier}
 #'  \item{USUBJID}{Unique Subject Identifier}
@@ -124,9 +124,6 @@
 #'  \item{AVALU}{Analysis Value - Units}
 #'  \item{ANL02FL}{Analysis Record Flag 02-Analysis Value}
 #'  \item{TRTEMFL}{Treatment Emergent Analysis Flag}
-#'  \item{COUNTRY_DECODE}{Country}
-#'  \item{RACE_DECODE}{Race Description}
-#'  \item{ETHNIC_DECODE}{Ethnicity Description}
 #'  \item{PARCAT2}{Parameter Category 2}
 #'  \item{PARCAT3}{Parameter Category 3}
 #'  \item{PARCAT4}{Parameter Category 4}
@@ -162,9 +159,9 @@
 #'  \item{ANL15FL}{Analysis Flag 15}
 #'  \item{ANL16FL}{Analysis Flag 16}
 #' }
-#' @seealso \code{\link{adae}} \code{\link{adaeocmq}} \code{\link{adagocmq}} \code{\link{adcm}} \code{\link{addili}} \code{\link{adeg}} \code{\link{adex}} \code{\link{adexsum}} \code{\link{adlb}} \code{\link{adpc}} \code{\link{adsl}} \code{\link{adttesaf}} \code{\link{advs}}# nolint
+#' @seealso \code{\link{adae}} \code{\link{adae.xpt}} \code{\link{adaeocmq}} \code{\link{adaeocmq.xpt}} \code{\link{adagocmq}} \code{\link{adagocmq.xpt}} \code{\link{adcm}} \code{\link{adcm.xpt}} \code{\link{addili}} \code{\link{addili.xpt}} \code{\link{adeg}} \code{\link{adeg.xpt}} \code{\link{adex}} \code{\link{adex.xpt}} \code{\link{adexsum}} \code{\link{adexsum.xpt}} \code{\link{adlb}} \code{\link{adlb.xpt}} \code{\link{adpc}} \code{\link{adpc.xpt}} \code{\link{adsl}} \code{\link{adsl.xpt}} \code{\link{adttesaf}} \code{\link{adttesaf.xpt}} \code{\link{advs}} \code{\link{advs.xpt}}# nolint
 #' @keywords datasets adlb
 #' @name adlb
 #' @examples
-#' head(data("adlb"))
+#'  head(data("adlb"))
 "adlb"