+++ data-raw/advs.R
@@ -441,8 +441,7 @@ gen_advs <- function(seed = 123) {
     "SAFFL",
     "STUDYID",
     "AGE",
-    "SEX",
-    "RACE_DECODE"
+    "SEX"
   )
 
   # Select only the key and the 'to_keep' variables from ADSL
@@ -461,7 +460,6 @@ gen_advs <- function(seed = 123) {
     AVALC = "Analysis Value (C)",
     AVALU = "Analysis Value Unit",
     ANL01FL = "Analysis Flag 01",
-    COUNTRY_DECODE = "Country",
     PARCAT1 = "Parameter Category 1",
     PARCAT2 = "Parameter Category 2",
     ADTM = "Analysis Date/Time",